Fix 'requireFileExistence' not using the correct constructor
The 'requireFileExistence' assertion function was not using the correct constructor before, as it was simply throwing in the given 'reason' string where the 'NoSuchFileException' was expecting a 'String' pointing towards the supposed file location, and not the actual error message. This has now been changed so that the thrown exception will be providing the correct information.

Signed-off-by: Oliver Berg <oliver@berg.moe>