[collections-0.11.0] Change how the 'Map<K, V>.getOrThrow' function works
Originally this function was naively assuming that 'V' could never be 'null' and just throwing an exception if the value stored under 'key' was 'null', this behaviour is not correct because depending on the implementation of the accessed 'Map' the value stored under 'key' may very well just be 'null', as 'null' is not explicitly used for signifying a missing value when invoking 'Map.get'.

The function will now only throw the exception if, and only if, the given 'key' is not contained within the accessed 'Map'.

Signed-off-by: Oliver Berg <oliver@berg.moe>