[collections-0.11.0] Deprecate 'Sequence.isEmpty' and 'Sequence.isNotEmpty' properties
These properties have been deprecated as they do not follow the established convention of using a function for 'isEmpty()' and 'isNotEmpty()' functions as set forth by the 'Collection' and 'Map' interfaces. And unlike the extension functions for 'Iterable' by the same name, these do not provide any chance for better performance than 'none()' and 'any()' as they only serve as wrappers for those functions.

Signed-off-by: Oliver Berg <oliver@berg.moe>