[collections-0.14.1] Deprecate 'enumSetOf' in favour of 'enumSetOfAll'
The reason for this deprecation is because the Kotlin std-lib, and also this library, use functions named 'XXXOf' with no parameters as a facade for 'emptyXXX' functions, which means that then having a function named 'enumSetOf' taking no parameters that instead fills the entire set, rather than creating an empty one, could lead to confusion, and as such it has been deprecated.

This function may be replaced with one that acts as a facade for 'emptyEnumSet' at some later point.

Signed-off-by: Oliver Berg <oliver@berg.moe>