[io-1.5.0] Deprecate setters that behave in a faulty manner
The setters that have been deprecated are; 'Path.name', 'Path.simpleName' and 'Path.extension'.
The reason for the deprecation is that the setters use the 'Path.renameTo' function, which is a wrapper for the 'Path.moveTo'/'Files.move' function, which in turn return the given 'target' as the return value, this is because when renaming the file we are just *moving* the file to a new path, meaning that after invoking the setter to say, 'Path.name', the value we *just* invoked that setter on will now be stale, as it does not exist anymore. This can cause some serious issues as it is not obvious at all that this is what happens, and therefore they have been deprecated.

Signed-off-by: Oliver Berg <oliver@berg.moe>